# Add latencies for 2xx responses