@chr1st1ank 2021-05-05
Hey @codesenberg,
any chance to get this PR into master? Or anything I can do to make it easier for you to review?
I tried to keep the changes minimal, but as the architecture was I had to touch a few files, unfortunately. A more general solution (e.g. a tabular output format with results on all status code ranges) I didn't try because this would need some more serious refactoring first.