@chr1st1ank 2022-01-06
As a preliminary solution I used my forked master branch to publish this feature: https://github.com/chr1st1ank/bombardier/releases/tag/v1.3.5

No offense meant. I just needed windows + linux builds of it on multiple machines. I'd be very happy to merge it into the main repository and get rid of my fork.